+++ tests/Console/ModelsCommand/Pivot/Models/Pivots/CustomPivot.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\Pivot\Models\Pivots;
+
+use Illuminate\Database\Eloquent\Relations\Pivot;
+
+class CustomPivot extends Pivot
+{
+
+}